start moving some tests into seprarte files, starting w/ the query transform processing
