+++ swirl/tests/tests.py
@@ -97,29 +97,6 @@ def test_utils_prefix_toks(prefix_toks_test_cases, prefix_toks_test_expected):
         i = i + 1
 
 
-class SearchTestCase(TestCase):
-    @pytest.fixture(autouse=True)
-    def _init_fixtures(self, api_client,test_suser, test_suser_pw):
-        self._api_client = api_client
-        self._test_suser = test_suser
-        self._test_suser_pw = test_suser_pw
-
-    def setUp(self):
-        settings.CELERY_TASK_ALWAYS_EAGER = True
-
-    def tearDown(self):
-        settings.CELERY_TASK_ALWAYS_EAGER = False
-
-    def test_query_search_viewset_search(self):
-        is_logged_in = self._api_client.login(username=self._test_suser.username, password=self._test_suser_pw)
-        # Check if the login was successful
-        assert is_logged_in, 'Client login failed'
-        # Call the viewset
-        surl = reverse('search')
-        response = self._api_client.get(surl, {'qs': 'notebook', 'pre_query_processor':'foo.bar'})
-
-
-
 @pytest.mark.django_db
 def test_query_transform_viewset_crud(api_client, test_suser, test_suser_pw, qrx_record_1):
